Migrate SectionProperty to separate file (read)
